Merge with tm1000/TableDnD as per pull request #7.
At #7 Andrew Nagy suggests the addition of a jsonize method.
I completely disagree with the implementation but he has a valid requirement.
Lets see if we can't fix it though and make everyone happy. =)

Added jsonize.
Which instead of the function serialize it will turn the ordered output into a valid json statement. This is useful for large tables since you can only send a specified URL length for each browser/server