Implement hierarchyDnD.
Configurable through the hierarchyLevel option, default set to 0 (off) which indicates the number of levels deep the hierarchy should be enabled.
By means of the direction detector it assesses whether you wish to move up or down the hierarchy.
The hierarchy placements left or right are facilitated through an invisible div classed "indent" and the nummber of div tags will indicate the current level of the rov in relation to the row before it.
Setting the first cell wide enough to let the indent divs, float left and specified width, move freely the first cel can comfortably indicate a hierarchy while the rest of the table data remains unmoved.
Naturally you can only move one level higher than the parent.
Refactored verticalDnD as well to keep things DRY.