Add mouse motion sensitivity adjustment.
The addition of a sensitivity value (default 10) which sets a threshold within which mouse motion will be ignored.
Any motion between + sensitivity and -sensitivity will not be considered as any movement.
This also has the side effect that it gives a much clearer indication of the actual direction the motion is taking.