Improve scope limiting closure.
Originally emplayed by facebook also used by twitter, a less error prone way to indicate the presence of a self executing closure. Instead of the () brackets we add a ! infront of the closure to indicate that this is a self execiting closure. Specificall finding juery in the glabal context ensures that we are nat dealing with a copy.